+++ io/src/lib.rs
@@ -1,6 +1,6 @@
 //! Async traits and types used for Io operations.
 
-#![deny(unsafe_code)]
+#![forbid(unsafe_code)]
 
 pub mod bytes;
 pub mod io;